+++ vakt/policy.py
@@ -3,6 +3,7 @@ Namespace for a basic Policy class.
 """
 
 import logging
+import warnings
 
 from .effects import ALLOW_ACCESS, DENY_ACCESS
 from .exceptions import PolicyCreationError
@@ -16,19 +17,29 @@ log = logging.getLogger(__name__)
 
 class Policy(JsonSerializer, PrettyPrint):
     """Represents a policy that regulates access and allowed actions of subjects
-    over some resources under a set of rules.
+    over some resources under a set of context restrictions.
     """
-    def __init__(self, uid, subjects=(), effect=DENY_ACCESS, resources=(), actions=(), rules=None, description=None):
+    def __init__(self, uid, subjects=(), effect=DENY_ACCESS, resources=(),
+                 actions=(), context=None, rules=None, description=None):
         self.uid = uid
         self.subjects = subjects
         self.effect = effect or DENY_ACCESS
         self.resources = resources
         self.actions = actions
-        rules = rules or {}
-        if not isinstance(rules, dict):
-            log.error('Error creating Policy. Rules must be a dictionary')
-            raise PolicyCreationError("Error creating Policy. Rules must be a dictionary")
-        self.rules = rules
+        # check for deprecated rules argument.
+        # If both 'context' and 'rules' are present - 'context' wins
+        if context:
+            pass
+        elif rules:
+            warnings.warn("'rules' argument will be removed in next version. Use 'context' argument",
+                          DeprecationWarning, stacklevel=2)
+            context = rules
+        else:
+            context = {}
+        if not isinstance(context, dict):
+            log.error('Error creating Policy. Context must be a dictionary')
+            raise PolicyCreationError("Error creating Policy. Context must be a dictionary")
+        self.context = context
         self.description = description
         self.type = self.get_type()
 
@@ -38,12 +49,17 @@ class Policy(JsonSerializer, PrettyPrint):
         if 'uid' not in props:
             log.error("Error creating policy from json. 'uid' attribute is required")
             raise PolicyCreationError("Error creating policy from json. 'uid' attribute is required")
-        rules = {}
-        if 'rules' in props:
-            for k, clazz in props['rules'].items():
-                rules[k] = Rule.from_json(clazz)
-        props['rules'] = rules
-        if 'type' in props:
+        context_rules = {}
+        if 'context' in props:
+            rules = props['context']
+        elif 'rules' in props:  # this is to support deprecated 'rules' attribute
+            rules = props['rules']
+        else:
+            rules = {}
+        for k, clazz in rules.items():
+            context_rules[k] = Rule.from_json(clazz)
+        props['context'] = context_rules
+        if 'type' in props:  # type is calculated dynamically on init
             del props['type']
         return cls(**props)